fix: resolve push blockers in durability hardening
Fix the hiring _store infinite recursion; update tests for the verify-full ssl_mode default, the distinct backup error codes, the new repo accessors on the protocol fakes, and the fail-closed code_runner sandbox wiring. Extract durability wiring, promotion persistence, and the engine max-turns resolver into siblings to stay under the module-size caps.
