+++ src/synthorg/hr/promotion/service.py
@@ -6,7 +6,6 @@ including criteria evaluation, approval decisions, model mapping,
 and trust integration.
 """
 
-import asyncio
 from collections.abc import Awaitable, Callable
 from datetime import UTC, datetime, timedelta
 
@@ -17,7 +16,6 @@ from synthorg.approval.protocol import ApprovalStoreProtocol
 from synthorg.core.agent import AgentIdentity
 from synthorg.core.concurrency.refcounted_lock_map import RefcountedLockMap
 from synthorg.core.critical_errors import reraise_critical
-from synthorg.core.persistence_errors import PersistenceError
 from synthorg.core.types import NotBlankStr
 from synthorg.hr.enums import PromotionDirection
 from synthorg.hr.errors import (
@@ -29,6 +27,7 @@ from synthorg.hr.performance.tracker import PerformanceTracker
 from synthorg.hr.promotion._approval_ops import create_approval, verify_approval
 from synthorg.hr.promotion._identity_guard import checked_identity
 from synthorg.hr.promotion._levels import next_level, prev_level
+from synthorg.hr.promotion._persistence import PromotionPersistenceMixin
 from synthorg.hr.promotion._record_builder import build_promotion_record
 from synthorg.hr.promotion.approval_protocol import PromotionApprovalStrategy
 from synthorg.hr.promotion.config import PromotionConfig
@@ -49,15 +48,12 @@ from synthorg.observability.events.promotion import (
     PROMOTION_EVALUATE_COMPLETE,
     PROMOTION_EVALUATE_FAILED,
     PROMOTION_EVALUATE_START,
-    PROMOTION_HISTORY_HYDRATED,
     PROMOTION_MODEL_CHANGED,
     PROMOTION_NOTIFICATION_SENT,
-    PROMOTION_PERSIST_FAILED,
     PROMOTION_REJECTED,
     PROMOTION_REQUESTED,
 )
 from synthorg.persistence.promotion_history_protocol import (
-    PromotionHistoryFilterSpec,
     PromotionHistoryRepository,
 )
 from synthorg.security.trust.service import TrustService
@@ -66,7 +62,6 @@ logger = get_logger(__name__)
 
 
 _SYSTEM_INITIATOR = NotBlankStr("system")
-_PERSIST_TIMEOUT_SECONDS: float = 5.0
 
 # Callback type for promotion/demotion notifications.
 # The communication layer can supply a concrete callback
@@ -77,7 +72,7 @@ PromotionNotificationCallback = Callable[
 ]
 
 
-class PromotionService:
+class PromotionService(PromotionPersistenceMixin):
     """Orchestrates agent promotions and demotions.
 
     Coordinates criteria evaluation, approval decisions, model
@@ -134,69 +129,6 @@ class PromotionService:
         # apply and double-promote. The lock makes the re-check-and-apply
         # atomic per agent; a different agent never contends.
         self._apply_locks = RefcountedLockMap[str]()
-
-    def attach_persistence(self, *, history_repo: PromotionHistoryRepository) -> None:
-        """Attach the durable promotion-history repo after boot.
-
-        The service is built in the construction phase before
-        persistence exists; this is called from the on-startup wiring
-        hook. Pair with :meth:`hydrate`.
-        """
-        self._history_repo = history_repo
-
-    async def hydrate(self) -> None:
-        """Load durable promotion history and recompute per-agent cooldown.
-
-        Idempotent and a no-op when no repository is attached. The
-        cooldown is derived from the newest record per agent so an
-        in-cooldown agent stays gated across a restart.
-        """
-        if self._history_repo is None:
-            return
-        history: dict[str, list[PromotionRecord]] = {}
-        offset = 0
-        page = 100
-        while True:
-            records = await self._history_repo.query(
-                PromotionHistoryFilterSpec(), limit=page, offset=offset
-            )
-            for record in records:
-                history.setdefault(str(record.agent_id), []).append(record)
-            if len(records) < page:
-                break
-            offset += page
-        cooldown: dict[str, AwareDatetime] = {}
-        for agent_key, records_list in history.items():
-            # query() returns newest-first; reverse to append (oldest-first)
-            # order and derive cooldown from the most-recent record.
-            newest = records_list[0]
-            records_list.reverse()
-            if self._config.cooldown_hours > 0:
-                cooldown[agent_key] = newest.effective_at + timedelta(
-                    hours=self._config.cooldown_hours
-                )
-        self._promotion_history = history
-        self._cooldown_until = cooldown
-        logger.info(
-            PROMOTION_HISTORY_HYDRATED,
-            agents=len(history),
-            records=sum(len(v) for v in history.values()),
-        )
-
-    async def _persist_record(self, record: PromotionRecord) -> None:
-        """Best-effort write-through of one promotion record."""
-        if self._history_repo is None:
-            return
-        try:
-            async with asyncio.timeout(_PERSIST_TIMEOUT_SECONDS):
-                await self._history_repo.append(record)
-        except (PersistenceError, TimeoutError) as exc:
-            logger.warning(
-                PROMOTION_PERSIST_FAILED,
-                agent_id=str(record.agent_id),
-                error_type=type(exc).__name__,
-                error=safe_error_description(exc),
-            )
 
     async def evaluate_promotion(
         self,
